+++ R/timezone.r
@@ -131,6 +131,84 @@ clean_tz <- function(tz, na = NA_character_) {
     tz
 }
 
+#' Change time zones while preserving UTC time
+#'
+#' `datetime_at_tz()` changes time zones while preserving UTC time (instead of clock time).
+#'
+#' @param x A datetime object.
+#' @param tz The target timezone to change to.
+#' @inheritParams as_sys_time.datetimeoffset
+#' @examples
+#' if(all(c("America/Los_Angeles", "America/New_York") %in% OlsonNames())) {
+#'   dt0 <- as_datetimeoffset("2020-01-01T01:01[America/Los_Angeles]")
+#'   dt <- datetime_at_tz(dt0, "America/New_York")
+#'   print(dt)
+#'   dt <- datetime_at_tz(as.POSIXct(dt0), "America/New_York")
+#'   print(dt)
+#'   dt <- datetime_at_tz(clock::as_zoned_time(dt0), "America/New_York")
+#'   print(dt)
+#'
+#'   # Can also use `lubridate::with_tz()`
+#'   if (require("lubridate")) {
+#'     dt <- lubridate::with_tz(dt0, "America/New_York")
+#'     print(dt)
+#'   }
+#' }
+#' @return A datetime object.  The UTC time should be the same but with a different time zone.
+#' @seealso [set_tz()] changes time zones while preserving clock time (instead of UTC time).
+#' @export
+datetime_at_tz <- function(x, tz = "", ...) {
+    UseMethod("datetime_at_tz")
+}
+
+#' @rdname datetime_at_tz
+#' @export
+datetime_at_tz.datetimeoffset <- function(x, tz = "", ...,
+                                             ambiguous = "error", nonexistent = "error",
+                                             fill = NA_character_) {
+    tz <- clean_tz(tz, na = NA_character_)
+    n <- max(length(x), length(tz))
+    if (length(x) < n)
+        x <- rep(x, length.out = n)
+    if (length(tz) < n)
+        tz <- rep(tz, length.out = n)
+    df <- data.frame(dt = x, tz = tz, stringsAsFactors = FALSE)
+    purrr::pmap_vec(df, with_tz_helper,
+                    ambiguous = ambiguous, nonexistent = nonexistent, fill = fill,
+                    .ptype = datetimeoffset())
+}
+
+#' @rdname datetime_at_tz
+#' @export
+datetime_at_tz.clock_zoned_time <- function(x, tz = "", ...) {
+    clock::zoned_time_set_zone(x, tz)
+}
+
+#' @rdname datetime_at_tz
+#' @export
+datetime_at_tz.POSIXt <- function(x, tz = "", ...) {
+    clock::date_set_zone(x, tz)
+}
+
+#' @rdname datetime_at_tz
+#' @export
+datetime_at_tz.default <- function(x, tz = "", ...) {
+    assert_suggested("lubridate")
+    lubridate::with_tz(x, tz)
+}
+
+with_tz_helper <- function(dt, tz, ambiguous = "error", nonexistent = "error", fill = NA_character_) {
+    st <- as_sys_time_dto(dt, ambiguous = ambiguous, nonexistent = nonexistent, fill = fill)
+    dto <- as_datetimeoffset.clock_zoned_time(clock::as_zoned_time(st, tz))
+    datetime_narrow.datetimeoffset(dto, datetime_precision.datetimeoffset(dt))
+}
+
+with_tz.datetimeoffset <- function(x, tzone = "", ...,
+                                   ambiguous = "error", nonexistent = "error", fill = NA_character_) {
+    datetime_at_tz.datetimeoffset(x, tzone,
+                                     ambiguous = ambiguous, nonexistent = nonexistent, fill = fill)
+}
+
 # is_valid_tz <- function(tz) {
 #     all(na_omit(tz) %in% OlsonNames())
 # }