+++ NEWS.md
@@ -81,7 +81,7 @@ Initial features
   + `minute()` and `minute()<-`
   + `second()` and `second()<-`
   + `date()` and `date()<-`
-  + `tz()`, `tz()<-`, and `force_tz()`
+  + `tz()`, `tz()<-`, `force_tz()`, and `with_tz()`
 
 * New accessor S3 methods:
 
@@ -102,6 +102,8 @@ Initial features
   + `is_datetimeoffset()` and `NA_datetimeoffset_`
   + `fill_tz()` and `fill_offsets()` fill in missing time zones and missing UTC offsets respectively.
   + `mode_tz()` is an S3 method that gets most common time zone for a datetime object
+  + `datetime_at_tz()` can be used to change the timezone 
+    (changes clock time, not system time)
   + `precision_to_int()` converts datetime precisions to an integer
   + Support for `{base}` datetime extractors `weekdays()`, `months()`, `quarters()`, and `julian()`
   + Support for `{lubridate}` datetime extractors `isoyear()`, `epiyear()`,